% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacharach-package.R
\docType{package}
\name{bacharach}
\alias{bacharach}
\alias{bacharach-package}
\title{Tools to Work with the Raindrop API}
\description{
Raindrop \url{https://raindrop.io} provides a URL bookmarking service
and an API for working with that service and more. Tools are provided to
work with the Raindrop API.nA Raindrop API key (free) is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/bacharach}
  \item Report bugs at \url{https://github.com/hrbrmstr/bacharach/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
