% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{raindrop_api_token}
\alias{raindrop_api_token}
\title{Get or set RAINDROP_API_TOKEN value}
\usage{
raindrop_api_token(force = FALSE)
}
\arguments{
\item{force}{Force setting a new Raindrop key for the current environment?}
}
\value{
atomic character vector containing the Raindrop token
}
\description{
The API wrapper functions in this package all rely on a Raindrop
token residing in the environment variable \code{RAINDROP_API_TOKEN}.
The easiest way to accomplish this is to set it
in the \code{.Renviron} file in your home directory.
}
\details{
You can obtain an Raindrop token by going visiting
\url{https://app.raindrop.io/#/settings/apps/dev}. Use the
\strong{test token}.
}
