% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{raindrop_parse_url}
\alias{raindrop_parse_url}
\title{Parse and extract useful info from any URL}
\usage{
raindrop_parse_url(url, api_key = raindrop_api_token())
}
\arguments{
\item{url}{a URL}
}
\description{
Parse and extract useful info from any URL
}
