% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacharach-package.R
\docType{package}
\name{bacharach}
\alias{bacharach}
\alias{bacharach-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/bacharach}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/bacharach/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
